move API DTOs to separate file
